requirements.txt
customtkinter
opencv-python
pyzbar
pandas
qrcode
Pillow
Flask
pandas
pathlib
requests
